acrescentados alguns endpoints ao código
